Add scuffed highlighting to 'vwf' command.
When this command opens a file, it will now attempt to highlight it. It does not yet do this based on file extension.

So files that are just text may be highlighted.

The plan is to add specific highlighters for specific file extensions. So '.cpp' will cause the highlighter to be a (for example) 'CPlusPlusHighligher'...

Small step to making builtin commands more fluent and polished.