OnSettingChange(): return plugin response
This also logs which plugin stopped the change, helping a user from stopping a rogue plugin if that happens.

GetPlugin<T>(): you can now get a plugin via its type. This is useful for commands that need to access their plugin. A simple `terminal.PluginManager.GetPlugin<YourType>()` is now possible.