Cross-Platform: Use different delimeters & SP's
The delimeter for the `PATH` variable is ';' on windows but ':' on GNU based systems (or macos), PathSep addresses this.

The search pattern for executables is different too. On windows, we check extensions because most CMD commands are just under System32.

On Linux however, directorys in the PATH variable usually only contains binarys. We have a simple check to make sure it has no extension to load it. This could be done better.